these don't have to be uint64

+++ agent/consul/config_endpoint_test.go
@@ -358,7 +358,7 @@ func TestConfigEntry_Get_BlockOnNonExistent(t *testing.T) {
 		})
 
 		g.Go(func() error {
-			for i := uint64(0); i < 200; i++ {
+			for i := 0; i < 200; i++ {
 				time.Sleep(5 * time.Millisecond)
 
 				var out bool
@@ -529,7 +529,7 @@ func TestConfigEntry_List_BlockOnNoChange(t *testing.T) {
 		})
 
 		g.Go(func() error {
-			for i := uint64(0); i < 200; i++ {
+			for i := 0; i < 200; i++ {
 				time.Sleep(5 * time.Millisecond)
 
 				var out bool
@@ -2215,7 +2215,7 @@ func TestConfigEntry_ResolveServiceConfig_BlockOnNoChange(t *testing.T) {
 		})
 
 		g.Go(func() error {
-			for i := uint64(0); i < 200; i++ {
+			for i := 0; i < 200; i++ {
 				time.Sleep(5 * time.Millisecond)
 
 				var out bool
